define API on request
